remove circular dependencies from bam track code
