+++ js/util/getChrColor.js
@@ -1,0 +1,75 @@
+import {IGVColor} from "../../node_modules/igv-utils/src/index.js"
+
+export function getChrColor(chr) {
+    if (chrColorMap[chr]) {
+        return chrColorMap[chr]
+    } else if (chrColorMap["chr" + chr]) {
+        const color = chrColorMap["chr" + chr]
+        chrColorMap[chr] = color
+        return color
+    } else {
+        const color = IGVColor.randomRGB(0, 255)
+        chrColorMap[chr] = color
+        return color
+    }
+}
+
+const chrColorMap = {
+    "chrX": "rgb(204, 153, 0)",
+    "chrY": "rgb(153, 204, 0)",
+    "chrUn": "rgb(50, 50, 50)",
+    "chr1": "rgb(80, 80, 255)",
+    "chrI": "rgb(139, 155, 187)",
+    "chr2": "rgb(206, 61, 50)",
+    "chrII": "rgb(206, 61, 50)",
+    "chr2a": "rgb(216, 71, 60)",
+    "chr2b": "rgb(226, 81, 70)",
+    "chr3": "rgb(116, 155, 88)",
+    "chrIII": "rgb(116, 155, 88)",
+    "chr4": "rgb(240, 230, 133)",
+    "chrIV": "rgb(240, 230, 133)",
+    "chr5": "rgb(70, 105, 131)",
+    "chr6": "rgb(186, 99, 56)",
+    "chr7": "rgb(93, 177, 221)",
+    "chr8": "rgb(128, 34, 104)",
+    "chr9": "rgb(107, 215, 107)",
+    "chr10": "rgb(213, 149, 167)",
+    "chr11": "rgb(146, 72, 34)",
+    "chr12": "rgb(131, 123, 141)",
+    "chr13": "rgb(199, 81, 39)",
+    "chr14": "rgb(213, 143, 92)",
+    "chr15": "rgb(122, 101, 165)",
+    "chr16": "rgb(228, 175, 105)",
+    "chr17": "rgb(59, 27, 83)",
+    "chr18": "rgb(205, 222, 183)",
+    "chr19": "rgb(97, 42, 121)",
+    "chr20": "rgb(174, 31, 99)",
+    "chr21": "rgb(231, 199, 111)",
+    "chr22": "rgb(90, 101, 94)",
+    "chr23": "rgb(204, 153, 0)",
+    "chr24": "rgb(153, 204, 0)",
+    "chr25": "rgb(51, 204, 0)",
+    "chr26": "rgb(0, 204, 51)",
+    "chr27": "rgb(0, 204, 153)",
+    "chr28": "rgb(0, 153, 204)",
+    "chr29": "rgb(10, 71, 255)",
+    "chr30": "rgb(71, 117, 255)",
+    "chr31": "rgb(255, 194, 10)",
+    "chr32": "rgb(255, 209, 71)",
+    "chr33": "rgb(153, 0, 51)",
+    "chr34": "rgb(153, 26, 0)",
+    "chr35": "rgb(153, 102, 0)",
+    "chr36": "rgb(128, 153, 0)",
+    "chr37": "rgb(51, 153, 0)",
+    "chr38": "rgb(0, 153, 26)",
+    "chr39": "rgb(0, 153, 102)",
+    "chr40": "rgb(0, 128, 153)",
+    "chr41": "rgb(0, 51, 153)",
+    "chr42": "rgb(26, 0, 153)",
+    "chr43": "rgb(102, 0, 153)",
+    "chr44": "rgb(153, 0, 128)",
+    "chr45": "rgb(214, 0, 71)",
+    "chr46": "rgb(255, 20, 99)",
+    "chr47": "rgb(0, 214, 143)",
+    "chr48": "rgb(20, 255, 177)",
+}
